Move some types from BuildTestBase.cs to separate files
- FileStat
- BuildPaths
- BlazorBuildOptions
- GlobalizationMode
