+++ src/mono/wasm/Wasm.Build.Tests/Common/BuildPaths.cs
@@ -1,0 +1,7 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#nullable enable
+
+namespace Wasm.Build.Tests;
+public record BuildPaths(string ObjWasmDir, string ObjDir, string BinDir, string BundleDir);
